backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "bcrypt==4.0.1  ",
    "fastapi>=0.135.0",
    "google-genai>=1.30.0",
    "passlib[bcrypt]>=1.7.4",
    "psycopg2-binary>=2.9.11",
    "pydantic-settings>=2.13.1",
    "pydantic[email]>=2.12.5",
    "pypdf>=6.7.4",
    "python-dotenv>=1.2.1",
    "python-jose>=3.5.0",
    "python-multipart>=0.0.22",
    "sqlalchemy>=2.0.48",
    "uvicorn>=0.41.0",
]
